+++ src/functional/otf/compile/build_systems/compiledb.py
@@ -6,7 +6,7 @@ import pathlib
 import subprocess
 from typing import Optional
 
-from functional.otf import languages, source, stages, step_types
+from functional.otf import languages, source, stages
 from functional.otf.compile import build_data, compiler
 from functional.otf.compile.build_systems import cmake, cmake_lists
 from functional.program_processors.builders import cache
@@ -14,7 +14,7 @@ from functional.program_processors.builders import cache
 
 @dataclasses.dataclass
 class CompiledbFactory(
-    step_types.BuildSystemProjectGenerator[
+    compiler.BuildSystemProjectGenerator[
         languages.Cpp, languages.LanguageWithHeaderFilesSettings, languages.Python
     ]
 ):
